# src/lib/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Prophezeiung API",
    "version": "1.0.0",
    "description": "API for the Prophezeiung prediction platform"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication using JWT"
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["USER", "ADMIN"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string",
                "nullable": true
              },
              "role": {
                "$ref": "#/components/schemas/Role"
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              },
              "avatarEffect": {
                "type": "string",
                "nullable": true
              },
              "avatarEffectColors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "id",
              "username",
              "displayName",
              "role",
              "avatarUrl",
              "avatarEffect",
              "avatarEffectColors"
            ]
          },
          "forcePasswordChange": {
            "type": "boolean"
          }
        },
        "required": ["user"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
      "PasswordLoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["username", "password"]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["success"]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": ["newPassword", "confirmPassword"]
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "prophecyId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "value", "prophecyId", "userId", "createdAt"]
      },
      "RateRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "minimum": -10,
            "maximum": 10
          }
        },
        "required": ["value"]
      },
...
      "UserStatus": {
        "type": "string",
        "enum": ["PENDING", "APPROVED", "SUSPENDED"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "avatarEffect": {
            "type": "string",
            "nullable": true
          },
          "avatarEffectColors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "prophecies": {
                "type": "number"
              },
              "ratings": {
                "type": "number"
              }
            },
            "required": ["prophecies", "ratings"]
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "avatarUrl",
          "avatarEffect",
          "avatarEffectColors",
          "role",
          "status",
          "createdAt"
        ]
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
    },
...
  },
  "paths": {
    "/api/auth/login/password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with username and password",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Account not approved or blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Change password for authenticated user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or wrong current password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prophecies/{id}/rate": {
      "post": {
        "tags": ["Prophecies"],
        "summary": "Rate a prophecy (-10 to +10)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Prophecy ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rating saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prophecy": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "averageRating": {
                          "type": "number",
                          "nullable": true
                        },
                        "ratingCount": {
                          "type": "number"
                        }
                      },
                      "required": ["id", "averageRating", "ratingCount"]
                    },
                    "rating": {
                      "$ref": "#/components/schemas/Rating"
                    }
                  },
                  "required": ["prophecy", "rating"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid rating or cannot rate own prophecy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prophecy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get user details (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": ["user"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "Update user status or role (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": ["user"]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete user (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete yourself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}